+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - `XIT BURN`: Add an `OVERALL` optional parameter to display overall burn only
 - `shipping-per-unit-price`: Remove the currency symbol in the `LMP` per-unit price label
+- Improve the sorting order of "drones" and "ship kits" categories
 
 ### Fixed